+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.3.3
+
+- fix: ensure serverToSocket handles sockets in strict sequence as they are
+  accepted
+- fix: ensure that, when one side is closed, all data received from that side 
+  has been delivered to the other side before closing the other side
+
 ## 2.3.2
 
 - fix: stability under load